* Fixed: Possible runtime assert
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@9015 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/multiselectdlg.cpp
@@ -75,7 +75,7 @@ void MultiSelectDlg::UpdateStatus()
 {
     int count = 0;
     wxCheckListBox* lst = XRCCTRL(*this, "lstItems", wxCheckListBox);
-    for (int i = 0; i < (int)lst->GetCount(); ++i)
+    for (unsigned int i = 0; i < lst->GetCount(); ++i)
     {
         if (lst->IsChecked(i))
             ++count;
@@ -89,7 +89,7 @@ wxArrayString MultiSelectDlg::GetSelectedStrings() const
 {
     wxArrayString ret;
     wxCheckListBox* lst = XRCCTRL(*this, "lstItems", wxCheckListBox);
-    for (int i = 0; i < (int)lst->GetCount(); ++i)
+    for (unsigned int i = 0; i < lst->GetCount(); ++i)
     {
         if (lst->IsChecked(i))
             ret.Add(lst->GetString(i));
@@ -101,7 +101,7 @@ wxArrayInt MultiSelectDlg::GetSelectedIndices() const
 {
     wxArrayInt ret;
     wxCheckListBox* lst = XRCCTRL(*this, "lstItems", wxCheckListBox);
-    for (int i = 0; i < (int)lst->GetCount(); ++i)
+    for (unsigned int i = 0; i < lst->GetCount(); ++i)
     {
         if (lst->IsChecked(i))
             ret.Add(i);
@@ -115,7 +115,7 @@ void MultiSelectDlg::SelectWildCard(const wxString& wild, bool select, bool clea
         return;
     wxArrayString wilds = GetArrayFromString(wild, _T(";"));
     wxCheckListBox* lst = XRCCTRL(*this, "lstItems", wxCheckListBox);
-    for (int i = 0; i < (int)lst->GetCount(); ++i)
+    for (unsigned int i = 0; i < lst->GetCount(); ++i)
     {
         if (clearOld || !lst->IsChecked(i))
         {
